+++ packages/graph-explorer/src/connector/sparql/sparqlExplorer.ts
@@ -1,6 +1,7 @@
 import type {
   Criterion,
   Explorer,
+  ExplorerRequestOptions,
   KeywordSearchRequest,
   KeywordSearchResponse,
   NeighborsResponse,
@@ -24,6 +25,7 @@ import { ConnectionConfig } from "@shared/types";
 import { v4 } from "uuid";
 import { env, logger } from "@/utils";
 import { createLoggerFromConnection } from "@/core/connector";
+import { FeatureFlags } from "@/core";
 
 const replaceBlankNodeFromSearch = (
   blankNodes: BlankNodesMap,
@@ -133,37 +135,49 @@ const storedBlankNodeNeighborsRequest = (
   });
 };
 
-function _sparqlFetch(connection: ConnectionConfig, options?: any) {
+function _sparqlFetch(
+  connection: ConnectionConfig,
+  featureFlags: FeatureFlags,
+  options?: ExplorerRequestOptions
+) {
   return async (queryTemplate: string) => {
     logger.debug(queryTemplate);
     const body = `query=${encodeURIComponent(queryTemplate)}`;
-    const headers =
-      options?.queryId && connection.proxyConnection === true
+    const queryId = options?.queryId;
+    const headers: Record<string, string> =
+      queryId && connection.proxyConnection === true
         ? {
             accept: "application/sparql-results+json",
             "Content-Type": "application/x-www-form-urlencoded",
-            queryId: options.queryId,
+            queryId: queryId,
           }
         : {
             accept: "application/sparql-results+json",
             "Content-Type": "application/x-www-form-urlencoded",
           };
-    return fetchDatabaseRequest(connection, `${connection.url}/sparql`, {
-      method: "POST",
-      headers,
-      body,
-      ...options,
-    });
+    return fetchDatabaseRequest(
+      connection,
+      featureFlags,
+      `${connection.url}/sparql`,
+      {
+        method: "POST",
+        headers,
+        body,
+        ...options,
+      }
+    );
   };
 }
 
 async function fetchSummary(
   connection: ConnectionConfig,
-  options: RequestInit
+  featureFlags: FeatureFlags,
+  options?: RequestInit
 ) {
   try {
     const response = await fetchDatabaseRequest(
       connection,
+      featureFlags,
       `${connection.url}/rdf/statistics/summary?mode=detailed`,
       {
         method: "GET",
@@ -181,6 +195,7 @@ async function fetchSummary(
 
 export function createSparqlExplorer(
   connection: ConnectionConfig,
+  featureFlags: FeatureFlags,
   blankNodes: BlankNodesMap
 ): Explorer {
   const remoteLogger = createLoggerFromConnection(connection);
@@ -188,12 +203,18 @@ export function createSparqlExplorer(
     connection: connection,
     async fetchSchema(options) {
       remoteLogger.info("[SPARQL Explorer] Fetching schema...");
-      const summary = await fetchSummary(connection, options);
-      return fetchSchema(_sparqlFetch(connection, options), summary);
+      const summary = await fetchSummary(connection, featureFlags, options);
+      return fetchSchema(
+        _sparqlFetch(connection, featureFlags, options),
+        summary
+      );
     },
     async fetchVertexCountsByType(req, options) {
       remoteLogger.info("[SPARQL Explorer] Fetching vertex counts by type...");
-      return fetchClassCounts(_sparqlFetch(connection, options), req);
+      return fetchClassCounts(
+        _sparqlFetch(connection, featureFlags, options),
+        req
+      );
     },
     async fetchNeighbors(req, options) {
       remoteLogger.info("[SPARQL Explorer] Fetching neighbors...");
@@ -215,7 +236,7 @@ export function createSparqlExplorer(
       }
 
       const response = await fetchNeighbors(
-        _sparqlFetch(connection, options),
+        _sparqlFetch(connection, featureFlags, options),
         request
       );
       const vertices = replaceBlankNodeFromNeighbors(
@@ -238,7 +259,7 @@ export function createSparqlExplorer(
 
       if (bNode && !bNode.neighbors) {
         const response = await fetchBlankNodeNeighbors(
-          _sparqlFetch(connection, options),
+          _sparqlFetch(connection, featureFlags, options),
           {
             resourceURI: bNode.vertex.data.id,
             resourceClass: bNode.vertex.data.type,
@@ -265,10 +286,13 @@ export function createSparqlExplorer(
         };
       }
 
-      return fetchNeighborsCount(_sparqlFetch(connection, options), {
-        resourceURI: req.vertexId,
-        limit: req.limit,
-      });
+      return fetchNeighborsCount(
+        _sparqlFetch(connection, featureFlags, options),
+        {
+          resourceURI: req.vertexId,
+          limit: req.limit,
+        }
+      );
     },
     async keywordSearch(req, options) {
       options ??= {};
@@ -286,7 +310,7 @@ export function createSparqlExplorer(
       };
 
       const response = await keywordSearch(
-        _sparqlFetch(connection, options),
+        _sparqlFetch(connection, featureFlags, options),
         reqParams
       );
       const vertices = replaceBlankNodeFromSearch(
@@ -297,5 +321,5 @@ export function createSparqlExplorer(
 
       return { vertices };
     },
-  };
+  } satisfies Explorer;
 }